fixed accidental crash from previous commit.

+++ TnTCheckpoint/Program.cs
@@ -368,11 +368,14 @@ namespace TnTCheckpoint
                     }
 
                     //672 570, 803 600
-                    if (CheckText("errorcode", ConvertAspectRatioCoords(35, 52.7777777), ConvertAspectRatioCoords(41.82291666, 55.555555555)))
+                    if (d2window.Left != d2window.Right & d2window.Top != d2window.Bottom & d2window.Left != d2window.Top & D2Process != null) //make sure the window actually exists so we dont error code.
                     {
-                        DiscordClient.Rest.SendMessageAsync(DiscordChannelID, "Unexpected error code. Restarting program to resolve.").Wait();
-                        //error code. kill game process and treat it like a crash.
-                        D2Process.Kill();
+                        if (CheckText("errorcode", ConvertAspectRatioCoords(35, 52.7777777), ConvertAspectRatioCoords(41.82291666, 55.555555555)))
+                        {
+                            DiscordClient.Rest.SendMessageAsync(DiscordChannelID, "Unexpected error code. Restarting program to resolve.").Wait();
+                            //error code. kill game process and treat it like a crash.
+                            D2Process.Kill();
+                        }
                     }
 
                     Thread.Sleep(500);
